# vmdb_database_settings: use actual db values