--- a/app/models/vmdb_database_setting.rb
+++ b/app/models/vmdb_database_setting.rb
@@ -4,10 +4,7 @@ class VmdbDatabaseSetting < ApplicationRecord
 
   virtual_belongs_to :vmdb_database
   virtual_column :description,      :type => :string
-  virtual_column :minimum_value,    :type => :integer
-  virtual_column :maximum_value,    :type => :integer
   virtual_column :vmdb_database_id, :type => :integer
-  virtual_column :value,            :type => :string
 
   attr_writer :vmdb_database
 
@@ -17,25 +14,13 @@ class VmdbDatabaseSetting < ApplicationRecord
 
   delegate :id, :to => :vmdb_database, :prefix => true
 
-  def minimum_value
-    min_val || ''
-  end
-
-  def maximum_value
-    max_val || ''
-  end
-
-  def value
-    setting || ''
-  end
+  alias_attribute :minimum_value, :min_val
+  alias_attribute :maximum_value, :max_val
+  alias_attribute :value, :setting
 
   def description
     desc = short_desc
     desc += "  #{extra_desc}" unless extra_desc.nil?
     desc
-  end
-
-  def unit
-    super || ''
   end
 end