If a value in the database is nil, use it

While fixing rbac scopes and this model, I wondered why we had these defaults.
@Fryguy any reason to default these strings to '' instead of nil?

Of note `alias_attribute` is a rails construct and `virtual_column` is ours