@chessbyte 2016-04-18
@Fryguy please review