@kbrock 2016-06-22
While I was not going for performance, here is the implications of going this route:

vmdb database settings page sorting by `minimum` or `maximum`

**before:**

|     ms |   ms- |queries | query (ms) |     rows |**before**
|    ---:|   ---:|  ---:|   ---:|      ---:| ---
|  229.3 |   0.0 |    7 |   7.6 |      244 |`/ops/db_list`
|  229.3 |  34.1 |      |       |          |`POST http://localhost:3000/ops/db_list`
|  195.1 |  41.2 |    2 |   0.5 |        1 |`.Executing action: db_list`
|    6.0 |   6.0 |    3 |   1.2 |        3 |`..ApplicationController#set_session_tenant`
|   52.3 |  36.3 |      |       |          |`..MiqReport#paged_view_search`
|   16.0 |   4.7 |    1 |   0.5 |        1 |`...Rbac.search`
|   11.2 |  11.1 |    1 |   5.4 |      239 |`....Rbac.find_targets_without_rbac`
|   78.7 |   7.0 |      |       |          |`..Rendering: layouts/_x_gtl`
|   16.9 |  15.3 |      |       |          |`..Rendering: layouts/_x_pagingcontrols`

**after:**

There is 1 more query: Before, every settings record was downloaded. Since the after case only downloads the records for the page, a `count(*)` is sent to determine total number of records.

|     ms |   ms- |queries | query (ms) |     rows |**after**
|    ---:|   ---:|  ---:|   ---:|      ---:| ---
|  161.2 |   0.0 |    8 |   4.4 |       25 |`/ops/db_list`
|  161.2 |  36.7 |      |       |          |`POST http://localhost:3000/ops/db_list`
|  124.6 |  44.9 |    2 |   0.6 |        1 |`.Executing action: db_list`
|    6.5 |   6.5 |    3 |   1.4 |        3 |`..ApplicationController#set_session_tenant`
|   14.6 |   5.9 |    1 |   1.2 |       20 |`..MiqReport#paged_view_search`
|    8.7 |   6.8 |    1 |   0.6 |        1 |`...Rbac.search`
|    1.8 |   1.7 |    1 |   0.6 |          |`....Rbac.find_targets_without_rbac`
|   44.5 |   7.0 |      |       |          |`..Rendering: layouts/_x_gtl`
|   14.2 |  13.0 |      |       |          |`..Rendering: layouts/_x_pagingcontrols`